pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-powered-data-science-assistant"
version = "0.1.0"
description = "An MCP server that turns an LLM into a conversational data analyst."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Farshad" }
]
license = { text = "MIT" }
keywords = ["mcp", "llm", "data-science", "machine-learning", "claude"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "mcp>=1.17.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "matplotlib>=3.8.0",
  "scikit-learn>=1.5.0",
  "requests>=2.32.0",
  "lxml>=5.2.0",
  "html5lib>=1.1",
  "python-dotenv>=1.0.1",
  "google-genai>=1.0.0",
  "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "mypy>=1.0.0",
  "black>=24.0.0",
  "ruff>=0.1.0",
]

[project.scripts]
mcp-ds-server = "mcp_data_science_assistant.server:main"
mcp-ds-chat = "mcp_data_science_assistant.chat_client:main"

[project.urls]
Homepage = "https://github.com/ieeexplorer/MCP-Powered-Data-Science-Assistant"
Repository = "https://github.com/ieeexplorer/MCP-Powered-Data-Science-Assistant"
Documentation = "https://github.com/ieeexplorer/MCP-Powered-Data-Science-Assistant/blob/main/docs/architecture.md"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
